ADT-9 SPOI-7897 #resolve #comment move aggregatorRegistry depended implementation from AppDataSingleSchemaDimensionStoreHDHT to AbstractAppDataDimensionStoreHDHT
