+++ library/src/main/java/com/datatorrent/lib/appdata/schemas/DimensionalConfigurationSchema.java
@@ -2178,10 +2178,21 @@ public class DimensionalConfigurationSchema
     return dimensionsDescriptorIDToIncrementalAggregatorIDs;
   }
 
+  /**
+   * this is the old interface, keep it as was.
+   *
+   * @return The dimensionsDescriptorIDToAggregatorIDs map.
+   */
+  public List<IntArrayList> getDimensionsDescriptorIDToAggregatorIDs()
+  {
+    return getDimensionsDescriptorIDToIncrementalAggregatorIDs();
+  }
+  
   public List<IntArrayList> getDimensionsDescriptorIDToCompositeAggregatorIDs()
   {
     return dimensionsDescriptorIDToCompositeAggregatorIDs;
   }
+
 
   /**
    * Returns the dimensionsDescriptorIDToKeys map.
